+++ app/App.tsx
@@ -7,24 +7,12 @@
 
 import React from 'react';
 import {NavigationContainer} from '@react-navigation/native';
-import {createNativeStackNavigator} from '@react-navigation/native-stack';
-import HomeScreen from './pages/HomeScreen';
-import AppleHealthTestScreen from './pages/AppleHealthTestScreen';
-import WebViewScreen from './pages/WebViewScreen';
-import {RootStackParamList} from './utils/routerType';
-import AsyncStorageScreenPage from './pages/AsyncStorageScreen';
-
-const Stack = createNativeStackNavigator<RootStackParamList>();
+import TestApp from './pages/test/TestApp';
 
 function App(): JSX.Element {
   return (
     <NavigationContainer>
-      <Stack.Navigator initialRouteName="Home">
-        <Stack.Screen name="Home" component={HomeScreen} />
-        <Stack.Screen name="Test" component={AppleHealthTestScreen} />
-        <Stack.Screen name="WebView" component={WebViewScreen} />
-        <Stack.Screen name="Storage" component={AsyncStorageScreenPage} />
-      </Stack.Navigator>
+      <TestApp />
     </NavigationContainer>
   );
 }
