+++ app/package.json
@@ -12,8 +12,10 @@
   },
   "dependencies": {
     "@react-native-async-storage/async-storage": "^1.23.1",
+    "@react-navigation/bottom-tabs": "^6.5.20",
     "@react-navigation/native": "^6.1.17",
     "@react-navigation/native-stack": "^6.9.26",
+    "@react-navigation/stack": "^6.3.29",
     "react": "18.2.0",
     "react-native": "0.73.7",
     "react-native-health": "^1.18.0",
