+++ app/utils/routerType.ts
@@ -1,6 +1,29 @@
-export type RootStackParamList = {
+import {BottomTabScreenProps} from '@react-navigation/bottom-tabs';
+import {
+  CompositeScreenProps,
+  NavigatorScreenParams,
+} from '@react-navigation/native';
+import type {StackScreenProps} from '@react-navigation/stack';
+
+export type TestAppRootStackParamList = {
+  BaroMeter: NavigatorScreenParams<HomeTabParamList>;
   Home: undefined;
   Test: undefined;
   WebView: {uri: string};
   Storage: {uri: string};
 };
+
+export type TestAppRootStackScreenProps<
+  T extends keyof TestAppRootStackParamList,
+> = StackScreenProps<TestAppRootStackParamList, T>;
+
+export type HomeTabParamList = {
+  Calendar: undefined;
+  Test: undefined;
+};
+
+export type HomeTabScreenProps<T extends keyof HomeTabParamList> =
+  CompositeScreenProps<
+    BottomTabScreenProps<HomeTabParamList, T>,
+    TestAppRootStackScreenProps<keyof TestAppRootStackParamList>
+  >;
