+++ package.json
@@ -2,43 +2,8 @@
   "name": "health-app",
   "version": "0.0.1",
   "private": true,
-  "scripts": {
-    "preinstall": "npx only-allow pnpm",
-    "android": "react-native run-android",
-    "ios": "react-native run-ios",
-    "lint": "eslint .",
-    "start": "react-native start",
-    "test": "jest"
-  },
-  "dependencies": {
-    "@react-native-async-storage/async-storage": "^1.21.0",
-    "@react-navigation/native": "^6.1.9",
-    "@react-navigation/native-stack": "^6.9.16",
-    "react": "18.2.0",
-    "react-native": "0.72.5",
-    "react-native-health": "^1.18.0",
-    "react-native-safe-area-context": "^4.7.4",
-    "react-native-screens": "^3.27.0",
-    "react-native-webview": "^13.6.2"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.20.0",
-    "@babel/preset-env": "^7.20.0",
-    "@babel/runtime": "^7.20.0",
-    "@react-native/eslint-config": "^0.72.2",
-    "@react-native/metro-config": "^0.72.11",
-    "@tsconfig/react-native": "^3.0.0",
-    "@types/react": "^18.0.24",
-    "@types/react-test-renderer": "^18.0.0",
-    "babel-jest": "^29.2.1",
-    "eslint": "^8.19.0",
-    "jest": "^29.2.1",
-    "metro-react-native-babel-preset": "0.76.8",
-    "prettier": "^2.4.1",
-    "react-test-renderer": "18.2.0",
-    "typescript": "4.8.4"
-  },
-  "engines": {
-    "node": ">=16"
-  }
+  "scripts": {},
+  "dependencies": {},
+  "devDependencies": {},
+  "engines": {}
 }
